opencv-test/requirements.txt
PySide6==6.5.3
opencv-contrib-python==4.8.0.76
Pillow==10.0.1
